requirements.txt
# Runtime requirements

pyserial
paho-mqtt
pymodbus

